requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-debug-toolbar==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
greenlet==3.3.0
idna==3.11
inflection==0.5.1
packaging==26.0
pillow==12.1.0
playwright==1.57.0
pyee==13.0.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
